+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- Support for getting a list of installable modules from pext.github.io
 - Busy indicator when the list of entries is empty for a more responsive look
 
 ## [0.2] - 2016-11-21