+++ CHANGELOG.md
@@ -10,12 +10,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - The entry and command list will now be emptied each time just before selection_made is called
 
 ### Added
-- Support for getting a list of installable modules from pext.github.io
 - Busy indicator when the list of entries is empty for a more responsive look
+- Support for getting a list of installable modules from pext.github.io
 
 ### Fixed
 - All commands now correctly show up after emptying the search bar
 - Module lists are now sorted alphabetically
+- selection_made is no longer unnecessarily triggered when closing the window
 
 ## [0.2] - 2016-11-21
 ### Added